requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
tqdm==4.67.1
urllib3==2.5.0
